--- Assets/Scripts/Unibas/DBIS/VREP/Core/ExhibitionManager.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a0e6239c12564e728acc8fcb7dc9885a
-timeCreated: 1546865715
